+++ src/index.ts
@@ -45,7 +45,15 @@ export type {
 } from './endpoints/brands/brandBySlug'
 
 // Shop Configuration
-export type { ShopConfigurationResponseData } from './endpoints/shopconfiguration/shopconfiguration'
+export type {
+  ShopConfiguration,
+  ShopCountryCustomData,
+  ShopCustomData,
+  /**
+   * @deprecated - will be removed in the next major version. Please switch to ShopConfiguration.
+   */
+  ShopConfiguration as ShopConfigurationResponseData,
+} from './types/ShopConfiguration'
 
 // Typeahead
 export type {
@@ -60,13 +68,23 @@ export type {
 
 // Wishlist
 export type {
+  Wishlist,
   WishlistItem,
-  WishlistResponseData,
+  WishlistItemGroup,
+  WishlistItemCustomData,
   WishlistWith,
-  GetWishlistParameters,
-} from './endpoints/wishlist/getWishlist'
+  /**
+   * @deprecated - will be removed in the next major version. Please switch to Wishlist.
+   */
+  Wishlist as WishlistResponseData,
+} from './types/Wishlist'
+export type { GetWishlistParameters } from './endpoints/wishlist/getWishlist'
+
 export type { DeleteWishlistParameters } from './endpoints/wishlist/deleteWishlistItem'
-export type { AddWishlistItemParameters } from './endpoints/wishlist/addWishlistItem'
+export type {
+  AddWishlistItemParameters,
+  AddToWishlistFailureKind,
+} from './endpoints/wishlist/addWishlistItem'
 
 // Campaigns
 export { CampaignSortOption } from './endpoints/campaigns/campaigns'
@@ -196,6 +214,8 @@ export type {
   PromotionCondition,
   PromotionCustomData,
   PromotionEffect,
+  BuyXGetYEffect,
+  AutomaticDiscountEffect,
 } from './types/Promotion'
 export type {
   PromotionsEndpointResponseData,